requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
Markdown==3.1.1
Owlready2==0.19
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
